add workflow for publishing articles
