+++ .github/workflows/article-publisher.yml
@@ -1,0 +1,34 @@
+name: Articles Publisher
+
+on:
+  pull_request:
+    types: [closed]
+
+jobs:
+  article-publisher:
+    if: github.event.pull_request.merged == true && contains(github.event.pull_request.title, 'post:')
+    runs-on: ubuntu-latest
+    steps:
+      - name: checkout
+        uses: actions/checkout@v3
+
+      - name: set up Node.js
+        uses: actions/setup-node@v3
+        with:
+          node-version: "20"
+
+      - name: install deps and build
+        run: |
+          cd ./publisher
+          npm install
+          npm run build
+
+      - name: create env
+        run: |
+          touch .env.publisher.local
+          echo HASHNODE_API_TOKEN=${{ secrets.HASHNODE_API_TOKEN }} >> .env
+          echo HASHNODE_PUBLICATION_ID=${{ secrets.HASHNODE_PUBLICATION_ID }} >> .env
+
+      - name: run publisher
+        run: |
+          node ./publish_articles.js ${{ github.event.pull_request.title }}
