+++ publisher/src/main.ts
@@ -4,10 +4,11 @@ import path from "path"
 import {Article, ArticleBuilder} from "./domain/article"
 import Publisher from "./domain/publisher"
 import HashnodePublisher from "./integrations/hashnode_publisher"
+import 'dotenv/config'
 
 interface HashNodeConfig {
-  api_token: string | null
-  publication_id: string | null
+  api_token?: string
+  publication_id?: string
 }
 
 interface Config {
@@ -16,7 +17,44 @@ interface Config {
   hashnode: HashNodeConfig | null
 }
 
-export function start(config: Config) {
+export function init(posts_directory: string) {
+  const args = process.argv.slice(2)
+
+  if (!args.length || args.length > 2) {
+    throw new Error("The arguments given are not correct")
+  }
+
+  if (args[0] == "post:") {
+    let post_title = args[1].trim()
+
+    if (!isValidPostTitle(post_title)) {
+      throw new Error("Post title is not valid")
+    }
+
+    let hashnode_token = process.env.HASHNODE_API_TOKEN
+    let hashnode_publication_id = process.env.HASHNODE_PUBLICATION_ID
+
+    let hashnode_enabled = hashnode_token !== undefined && hashnode_publication_id !== undefined
+
+    let config: Config = {
+      post_title,
+      posts_directory,
+      hashnode: null,
+    }
+
+    if (hashnode_enabled) {
+      let hashnode_config: HashNodeConfig = {
+        api_token: hashnode_token,
+        publication_id: hashnode_publication_id,
+      }
+      config["hashnode"] = hashnode_config
+    }
+
+    start_publishing(config)
+  }
+}
+
+function start_publishing(config: Config) {
   if (!config.post_title) {
     throw new Error("Must provide the title of the blog")
   }
@@ -40,7 +78,6 @@ export function start(config: Config) {
     .build()
 
   const publishers: Publisher[] = []
-
   if (config.hashnode) {
     let {api_token, publication_id} = config.hashnode
 
@@ -58,4 +95,9 @@ export function start(config: Config) {
   for (let publisher of publishers) {
     publisher.publish(article)
   }
+}
+
+function isValidPostTitle(input: string) {
+  const pattern = /^[a-z](-?[a-z])*$/
+  return pattern.test(input)
 }
